refactor: move the sharedpref into the app-series module
Since the SharedPref class is currently only used in the app-series, move it into there
