+++ core/build.gradle
@@ -50,7 +50,6 @@ dependencies {
     implementation "com.google.android.material:material:$material_version"
     implementation "com.google.dagger:dagger:$dagger_version"
     implementation "com.google.dagger:dagger-android:$dagger_version"
-    implementation "com.squareup.moshi:moshi:$moshi_version"
 
     testImplementation 'io.mockk:mockk:1.9.3'
     testImplementation 'junit:junit:4.13'
